backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
base58==2.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
ECPy==1.2.5
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
types-Deprecated==1.2.15.20250304
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
xrpl-py==4.2.0
